+++ src/App.tsx
@@ -3,8 +3,10 @@
 import React, { useState, useEffect } from 'react';
 import { UserWarning } from './UserWarning';
 import { USER_ID } from './api/todos';
-import classNames from 'classnames';
 import { ErrorNotification } from './ErrorNotification';
+import { Header } from './Header';
+import { TodoList } from './TodoList';
+import { Footer } from './Footer';
 
 import { getTodos } from '../src/api/todos';
 import { Todo } from '../src/types/Todo';
@@ -62,116 +64,17 @@ export const App: React.FC = () => {
     return <UserWarning />;
   }
 
-  const filteredTodos = todos.filter(todo => {
-    if (filter === 'active') return !todo.completed;
-    if (filter === 'completed') return todo.completed;
-    return true;
-  });
-
-  const activeTodosCount = todos.filter(todo => !todo.completed).length;
-
   return (
     <div className="todoapp">
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {/* this button should have `active` class only if all todos are completed */}
-          <button
-            type="button"
-            className="todoapp__toggle-all active"
-            data-cy="ToggleAllButton"
-          />
-
-          {/* Add a todo on form submit */}
-          <form>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-            />
-          </form>
-        </header>
-
-        <section className="todoapp__main" data-cy="TodoList">
-
-          {filteredTodos.map(todo => (
-          <div data-cy="Todo" key={todo.id} className={classNames('todo', { completed: todo.completed })}>
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-                checked={todo.completed}
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              {todo.title}
-            </span>
-
-            {/* Remove button appears only on hover */}
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              ×
-            </button>
-
-            {/* overlay will cover the todo while it is being deleted or updated */}
-            <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div>
-          ))}
-        </section>
-
+        <Header />
+        <TodoList todos={todos} setTodos={setTodos} filter={filter}/>
         {/* Hide the footer if there are no todos */}
         {todos.length > 0 && (
 
-        <footer className="todoapp__footer" data-cy="Footer">
-          <span className="todo-count" data-cy="TodosCounter">
-            {activeTodosCount} items left
-          </span>
-
-          {/* Active link should have the 'selected' class */}
-          <nav className="filter" data-cy="Filter">
-            <a
-              href="#/"
-              className={`filter__link ${filter === 'all' ? 'selected' : ''}`}
-              onClick={() => setFilter('all')}
-              data-cy="FilterLinkAll"
-            >
-              All
-            </a>
-
-            <a
-              href="#/active"
-              className={`filter__link ${filter === 'active' ? 'selected' : ''}`}
-              onClick={() => setFilter('active')}
-              data-cy="FilterLinkActive"
-            >
-              Active
-            </a>
-
-            <a
-              href="#/completed"
-              className={`filter__link ${filter === 'completed' ? 'selected' : ''}`}
-              onClick={() => setFilter('completed')}
-              data-cy="FilterLinkCompleted"
-            >
-              Completed
-            </a>
-          </nav>
-
-          {/* this button should be disabled if there are no completed todos */}
-          <button
-            type="button"
-            className="todoapp__clear-completed"
-            data-cy="ClearCompletedButton"
-          >
-            Clear completed
-          </button>
-        </footer>
+        <Footer todos={todos} filter={filter} setFilter={setFilter}/>
         )}
       </div>
       <ErrorNotification
