+++ src/Footer.tsx
@@ -1,0 +1,60 @@
+import React from 'react';
+import { Todo } from './types/Todo';
+
+interface FooterProps {
+  todos: Todo[];
+  filter: 'all' | 'active' | 'completed';
+  setFilter: React.Dispatch<React.SetStateAction<'all' | 'active' | 'completed'>>;
+}
+
+export const Footer: React.FC<FooterProps> = ({ todos, filter, setFilter}) => {
+
+  const activeTodosCount = todos.filter(todo => !todo.completed).length;
+
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+          <span className="todo-count" data-cy="TodosCounter">
+            {activeTodosCount} items left
+          </span>
+
+          {/* Active link should have the 'selected' class */}
+          <nav className="filter" data-cy="Filter">
+            <a
+              href="#/"
+              className={`filter__link ${filter === 'all' ? 'selected' : ''}`}
+              onClick={() => setFilter('all')}
+              data-cy="FilterLinkAll"
+            >
+              All
+            </a>
+
+            <a
+              href="#/active"
+              className={`filter__link ${filter === 'active' ? 'selected' : ''}`}
+              onClick={() => setFilter('active')}
+              data-cy="FilterLinkActive"
+            >
+              Active
+            </a>
+
+            <a
+              href="#/completed"
+              className={`filter__link ${filter === 'completed' ? 'selected' : ''}`}
+              onClick={() => setFilter('completed')}
+              data-cy="FilterLinkCompleted"
+            >
+              Completed
+            </a>
+          </nav>
+
+          {/* this button should be disabled if there are no completed todos */}
+          <button
+            type="button"
+            className="todoapp__clear-completed"
+            data-cy="ClearCompletedButton"
+          >
+            Clear completed
+          </button>
+        </footer>
+  )
+}
