+++ src/TodoItem.tsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import classNames from 'classnames';
+import { Todo } from './types/Todo';
+
+interface TodoItemProps {
+  todo: Todo;
+}
+
+export const TodoItem: React.FC<TodoItemProps> = ({ todo }) => {
+
+  return (
+    <div data-cy="Todo" key={todo.id} className={classNames('todo', { completed: todo.completed })}>
+        <label className="todo__status-label">
+          <input
+            data-cy="TodoStatus"
+            type="checkbox"
+            className="todo__status"
+            checked={todo.completed}
+          />
+        </label>
+
+        <span data-cy="TodoTitle" className="todo__title">
+          {todo.title}
+        </span>
+
+        {/* Remove button appears only on hover */}
+        <button type="button" className="todo__remove" data-cy="TodoDelete">
+          ×
+        </button>
+
+        {/* overlay will cover the todo while it is being deleted or updated */}
+        <div data-cy="TodoLoader" className="modal overlay">
+          <div className="modal-background has-background-white-ter" />
+          <div className="loader" />
+        </div>
+      </div>
+  )
+}
